Updated event component to allow for new dialog. Code cleanup
